+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+[met4j-toolbox] Add Apps to compute compound graph's classical weights (degree/chemical similarity)
 [met4j-graph] Add method to create RPAIRs-like tags on compound graph's edges